Reproduce cookie refresh bugs, validate fixes
Per bitly/oauth2_proxy#115, @jehiah noted that he encountered a cookie refresh
bug whereby an empty cookie value was being set when --cookie-refresh was
enabled. I was able to ascertain by studying the code that if the original
cookie has expired, the "refresh" case would still be triggered, causing an
empty cookie to be set.

TestProcessCookieFailIfRefreshSetAndCookieExpired reproduces the bug when the
`ok &&` component of the `ok && p.CookieRefresh != time.Duration(0)`
expression within ProcessCookie() is removed. The bug is confirmed fixed when
the `ok &&` component is reinstated.

At the same time, it's now apparent that the effective cookie expiration
period was always one week prior to @jehiah's changes, thanks to the timestamp
in validateCookie() being compared against the hardcoded value
`time.Duration(24)*7*time.Hour*-1`. TestProcessCookieFailIfCookieExpired
confirms that @jehiah's changes solves this problem as well.

I also took the liberty to add a comment to the timestamp comparison in
validateCookie(), since it took effort to remind myself how it was supposed to
work.